Consecutive time signatures not properly merged #153
This was due to time signatures not being merged before getBarTimeSignatures/getBarDurations step.
In the example case

	timeSignature (6/4) $ timeSignatureDuring (1 <-> 2) (5/4) $ times 10 $ scat [c..c']^/8

what happens is actually that time signatures are represented as [(1, 6/4), (1, 6/4), (8, 6/4)] to which
getBarTimeSignature correctly responds [4/4,4/4,6/4..], shortening the first two bars because of the
subsequent time signature change.

Note that we still need retainUpdates after generating the time signature list to prevent redundant
time signature notations.

Closes #153