Add isTieBeginning/isTieEnd
Needed to be able to merge tied notes, see #137

This scheme works for everything except chords as there is no unambigous way
to say whether a score is tied or not (only some of the notes may be tied).

Hence the following is undefined

>>> isTieEndBeginning (c :: StandardNote)
undefined

Not sure how to solve this.